+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/observable/Retry.kt
@@ -12,7 +12,7 @@ import com.badoo.reaktive.disposable.Disposable
  *
  * Please refer to the corresponding RxJava [document](http://reactivex.io/RxJava/javadoc/io/reactivex/Observable.html#retry-io.reactivex.functions.BiPredicate-).
  */
-fun <T> Observable<T>.retry(predicate: (attempt: Int, Throwable) -> Boolean = { _, _ -> true }): Observable<T> =
+fun <T> Observable<T>.retry(predicate: (attempt: Long, Throwable) -> Boolean = { _, _ -> true }): Observable<T> =
     observable { emitter ->
         subscribe(
             object : ObservableObserver<T>, ValueCallback<T> by emitter, CompleteCallback by emitter {
